Update init.lua
Incorrect usage of logical operators in the code provided needed to be fixed. ~= is used for comparing two values and making sure they are NOT the same. However, there is usage of ~(conditional) and !conditional in this code, which is not valid lua. This has been replaced with "not <conditional>" as is proper.